Item B-03526: As a Manager, I wold like to sort services in the service management table
Add sorting and filter support to services in the same way it is done for ideas.

The services html is using a different design than the other similar html files, such as ideas.
Update services html to function more similar to that of the other html files.

This is not yet complete and is being pushed due to a request.