Defect D-01186: Non-Public Services showing for anonymous user
Add failsafe as public view if user role is undefined.
Weaver-UI-Core currently does not define "role" if the user is anonymous.

Non-public accounts may not have the role property assigned.
If the role is undefined, assume user is anonymous.

There is a setting 'anonymousRole' in app/config/appConfig.js.
I cannot find anywhere this is being used.

Ideally, this would need to respect the 'anonymousRole' appConfig.js setting but the solution may need to be done in Weaver-UI-Core instead of here.
This maintains the existing check for ROLE_ANONYMOUS such that if and when Weaver-UI-Core makes the mentioned change, this file need not be altered.