Item b03526: Present tabs based on roles, using ng-show
Perform access control in the controller and present a list of columns for renderring such than an ng-repeat can be used:
  `<wvr-tab ng-repeat="column in columns" path="{{column.path}}" view="views/management/{{column.view}}.html" label="{{column.label}}">{{column.label}}</wvr-tab>`

This is not working as expected, possibly due to the current weaver tabs implementation.
This includes a not so great work-around using ng-show with a preset number of columns.

This should be improved to use ng-repeat or something similar instead of using ng-show with a preset column list.