Fix regression in D-01186, introduced by B-03523
The solution for B-03523 incorrectly tests when the status is set to "DOWN".
It appears to be testings if the service is visible to the public OR if the status is "DOWN".
This ends up exposing non-public services to anonymous, thereby re-introducing D-01186.

Change the filterByStatus() condition parameter to an array so that multiple status can be checked via a single pass.