Implement FeatureProposalState on the UI
To perform a test against ng-if="" against the state, this adds a function hasState() to the controller.
Doing this seems to prevent any race conditions where FeatureProposalState is somehow not defined when used directly inside of an ng-if.

B-03632 implies some text on the states.
Add the text to the state and provide the state text as a summary description tooltip for the presented state.