Allow sorting of feature proposals by utilizing weaver-table
There already is a weaverTable columns variable defined, so create a variable called weaverTableDetail.
Setting `weaver-table="weaverTable"` to `weaver-table="weaverTableDetail"` does not work entirely as expected because angular templateUrl does not seem to inherit the modified scope.
As a work-around until a more appropriate solution can be implemented, provide a function to overwrite the weaverTable scope variable in the controller.