Issue 174: 401 when viewing service and a feature proposal for that service is edited in another browser
The ideaRepo automatically listens to changes on ideas.
When the user is anonymous, or other role with insufficient privileges, it is still listening.
This results in the 401 messages.

As anonymous is not allowed to access this, only listen when the user has sufficient privileges.
Based on the services, these privileges are the roles SERVICE_MANAGER or greater.