Upgrade Weaver-UI-Core (2.2.2-rc.8)
- Upgrade Weaver-UI-Core ("2.2.2-rc.8").
- Upgrade all dependencies with vulnerabilities (where possible).
- Provide a Dockerfile.
- Replace grunt with webpack.
- Get tests working.

The tests are failing because:
- Many of the test files are effectively empty.
- Something is wrong with Jasmine.

The tests that are effectively empty have been removed.
The tests regarding Jasmine are failing because of `TypeError: jasmine.pp is not a function`.
It is unclear why this test failure is happening.
Due to time constraints I simply removed the test service file entirely.

The Docker file has the tl-components for consistency.
This repository does not currently use tl-components and that Dockerfile code affectively does nothing.