Utilize `to()` method to custom craft paths.
The paths tinymce are looking for end up using only the immediate parent directory.
The design of the default `to` behavior does not support this.
Use the `to` function to manually craft the URL.

This handles most of the cases but there is still a problem with the theme.
There is a `modern` theme being used that does note exist.